add Makefile and rm variable length arrays

+++ RequestHandler.cpp
@@ -15,9 +15,9 @@ void	RequestHandler::handleRequest(struct kevent event_lst_item)
 	char buf[BUFFER_SIZE];
 	if (event_lst_item.filter == EVFILT_READ)
 	{
-		size_t bytes_read = recv(event_lst_item.ident, buf, sizeof(buf), 0);
+		int bytes_read = recv(event_lst_item.ident, buf, sizeof(buf), 0);
 		if (bytes_read == -1)
 			throw CustomException("Failed when processing read request\n");
-		printf("read %zu bytes\n", bytes_read);	
+		printf("read %i bytes\n", bytes_read);	
 	}
 }
